requirements.txt
langchain_openai
langchain_core
python-dotenv
streamlit
langchain_community
fastapi
langserve
uvicorn
sse_starlette